Refacotr controllers

+++ WSContabilidad/Models/Asiento.cs
@@ -13,10 +13,10 @@ namespace WSContabilidad.Models
         public DateTime Fecha { get; set; }
         public string Estado { get; set; }
         public int Moneda { get; set; }
-        public float TasaCambio { get; set; }
+        public double TasaCambio { get; set; }
         public int Cuenta { get; set; }
         public int TipoMovimiento { get; set; }
-        public float Monto { get; set; }
+        public double Monto { get; set; }
 
         //public CatalogoAuxiliar CatalogoAuxiliar { get; set; }
         //public Monedas Moneda { get; set; }
